Secure password input for login
Just like the Register form, the Login was not correctly handling
the user input for the password field. The entered password was
visible to anyone passing by or whatnot. Therefore 'secureTextEntry'
was also set to true for the password textinput in the Login form
as it was done in the Register form.