Change all instances of name to username
In the database setup in the back-end via a rails app the User
model was instantiated with the attribute 'username' instead of
'name', which was causing problems with out front-end since all
instances of calling for user information was requesting 'name'.
Therefore all back-end references to a user's name was changed to
'username'.