Add access token and redirect
The login page was not correctly saving users or redirecting so
access token methods were added to the login as well as a redirect
for when the login is successful. However, the login currently still
does not successfully save the state or current user of the app. That
feature/process will be added in the morning after watching a new
video on adding states to your app.
s