Have register persist user in database
Our previous version of Register was not functioning correctly
because the fetch was going to the wrong route from our back-end.
This was simply fixed by changing the route specified in the fetch.
Register also creates and saves an access token just like Login does,
but also doesn't save the state of the app and will be fixed in the
morning. Register does redirect to the Featured page upon completion.