Reformat Register fetch call
The previous fetch call was only handling part of what we needed
and was not setup for persisting a user throughout the app, therefore
a lot of code was added to help with access tokens and persisting
a logged in user. A catch call was also added so that any form errors
could be caught and displayed on the screen for the user to correct
their mistakes and actually be added as a new user. Currently the
registration does not fully work since a lot of our back-end for
users and sessions needs to be changed to go with this process.