Make BottomTabBar its own separate component file
The index page contained the TabBar component/view which meant that
automatically any person to visit the app would see inside the app
without having to login/register and that all of the tabs were on
the main page. Since this is not how we wanted the app to work the
TabBar section was moved to its own page and created as its own
component that could be implemented in the app after a user was
created/logged in.

- Make TabBar section its own component to be viewed later