Minor syntax changes
The redirect action of the login was erroring out because it was
set to go to 'home' which was not a component/page that we had set
up so it was changed to 'bottomtabbar' which contains of the inside
features of the app. A removeToken() line was also added in the
case of a caught error. Everything else was moved down one line on accident