Add section to print form errors
Originally the app was catching the errors caused by incorrectly
filling out the form, but was not rendering them on the page, and
would still go on to the next page. The app should catch the errors
and print them on the screen below the form so that the user knows
what they entered was incorrect and how to correct it. It should also
not advance to the inside of the app without the user being properly
registered, because this would defeat the whole purpose of having
users register to use the app. Therefore an errors section was added
to the view rendering to display these messages and a throw errors
line was added to the onRegisterPressed() function.