TabBar functionality moved to separate page, nav added
The TabBar component/section was showing up on the main page of the
app with all of the exclusive app content without having a user
login or register, so that functionality was moved to separate page
and is only called upon once a user is logged in. To make this work
a navigation section was added in place of the TabBar components
on the index page, which is what calls the main page, login, register,
and finally the BottomTabBar component which is the handling for the
entire rest of the app.