Add opening welcome page
Our app was opening and going straight to a login or register page,
which is not very inviting for a user. It also limited you to just
that page instead of choosing login or register, which is why the
WelcomePage was created. It is set at the first screen the user
sees when they open up the app and it includes both options of
either login or register.