Change bookfinder to Connected
Our app is called Connected and not bookfinder, but the app was
originally named that from a tutorial that we continued to build
our project off of. However, we no longer wanted this to show and
wanted our app's actual name to rendered as the name of the app
and project. To do so I had to modify the LaunchScreen to include
'Connected' and not 'bookfinder'. All of the files created
specifically for 'bookfinder' were also deleted to clean out the
log of the app and minimize confusion.