Move 'move' icon to far right
The app was displaying the 'more' tab in the middle of all the icons
which looks very dumb from a user and logical reasoning standpoint.
Generally you list all the specifics THEN put the more section,
which is why the 'more' tab was moved to the far right to come after
all the other icons in a conventional viewing method.