Change contacts tab to user profile tab
There was a tab in the bottom nav bar that linked to a contacts page
but we have no need or want for that in our app, so it was changed
to be a tab for the current user's profile page. The tab goes to
the profile page of the logged in user where they can see all of their
information. (possibly edit later)